lab6/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.26.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask-restful==0.3.10
flasgger==0.9.7.1
